Support TR::Block in TR_RegionStructure::cleanupAfterNodeRemoval()
This commit fixes an edge case in
TR_RegionStructure::cleanupAfterNodeRemoval() discovered by
eclipse-openj9/openj9#17419. Previously, it was expected that
a region structure could only collapsed into its parent iff
the remaining sub node was also a structure.
This did not account for the case that the sub node could
be a block.

Signed-off-by: James You <james.you@protonmail.com>