Add check for static class objects in Node::mayModifyValue
This commit fixes a segmentation fault that occurs during
side effect detection in IL generation for an edge case
a static class object reference is passed in as an argument
to Node::mayModifyValue.

Node::mayModifyValue assumes that if its opcode is a store,
the symbol reference whose value it modifies must be one of:

+ auto
+ parm
+ static field ref

Because of the way that side effects of nodes are
checked by recursively traversing their children
(see J9ByteCodeIlGenerator::valueMayBeModified)
and the way certain IL is generated,
it is possible that these assumptions are not invariant.

In a concrete case, when a awrtbar is generated for static stores,
one of the awrtbar children will be a loadaddr which loads
a static class object. When this case occurs, node::mayModifyValue
will erroneously use the constant pool index of the static class reference
as a static field reference, eventually resulting in a garbage value
and causing a segmentation fault.

Fixes: #18156

Signed-off-by: James You <james.you@protonmail.com>